Fix crash in RawBinaryReader when using large SkipHeaderBytes values
The issue appeared because the assumption was that the amount of data eventually
read would be larger than the number of header bytes to skip. Implemented a UInt64
filter parameter and matching widget. This new filter parameter is used for the
RawBinaryReader so that the use can have offsets larger than 2GB (2^31). The original
value was stored in an int32_t which would eventually be inadequate.

#fixes #329. closes #329. updates #329

+++ Source/SIMPLib/CoreFilters/RawBinaryReader.h
@@ -50,7 +50,7 @@ class SIMPLib_EXPORT RawBinaryReader : public AbstractFilter
     PYB11_PROPERTY(SIMPL::NumericTypes::Type ScalarType READ getScalarType WRITE setScalarType)
     PYB11_PROPERTY(int Endian READ getEndian WRITE setEndian)
     PYB11_PROPERTY(int NumberOfComponents READ getNumberOfComponents WRITE setNumberOfComponents)
-    PYB11_PROPERTY(int SkipHeaderBytes READ getSkipHeaderBytes WRITE setSkipHeaderBytes)
+    PYB11_PROPERTY(uint64_t SkipHeaderBytes READ getSkipHeaderBytes WRITE setSkipHeaderBytes)
     PYB11_PROPERTY(QString InputFile READ getInputFile WRITE setInputFile)
 
   public:
@@ -72,8 +72,8 @@ class SIMPLib_EXPORT RawBinaryReader : public AbstractFilter
     SIMPL_FILTER_PARAMETER(int, NumberOfComponents)
     Q_PROPERTY(int NumberOfComponents READ getNumberOfComponents WRITE setNumberOfComponents)
 
-    SIMPL_FILTER_PARAMETER(int, SkipHeaderBytes)
-    Q_PROPERTY(int SkipHeaderBytes READ getSkipHeaderBytes WRITE setSkipHeaderBytes)
+    SIMPL_FILTER_PARAMETER(uint64_t, SkipHeaderBytes)
+    Q_PROPERTY(uint64_t SkipHeaderBytes READ getSkipHeaderBytes WRITE setSkipHeaderBytes)
 
     SIMPL_FILTER_PARAMETER(QString, InputFile)
     Q_PROPERTY(QString InputFile READ getInputFile WRITE setInputFile)
